freeze migrations into v10.11.11_0.ts, bump to 10.11.11:1
